+++ Tests/Utilities/DisplayNameHelperTests.cs
@@ -7,15 +7,16 @@ namespace Macabresoft.Core.Tests.Utilities {
     [TestFixture]
     public static class Tests {
         private const string ClassName = "All These Governors";
-
+        private const string Property1Name = "The Hero of Kvatch!";
+        private const string Property2Name = "Tarhiel";
         [Test]
         [Category("Unit Tests")]
         public static void GetEnumDisplayName_Should_UseDisplayAttribute() {
             using (new AssertionScope()) {
-                DisplayNameHelper.GetEnumDisplayName(TestEnum.ValueWithName1).Should().Be(TestEnumNames.Name1);
-                DisplayNameHelper.GetEnumDisplayName(TestEnum.ValueWithName2).Should().Be(TestEnumNames.Name2);
-                DisplayNameHelper.GetEnumDisplayName(TestEnum.ValueWithoutName1).Should().Be(TestEnum.ValueWithoutName1.ToString());
-                DisplayNameHelper.GetEnumDisplayName(TestEnum.ValueWithoutName2).Should().Be(TestEnum.ValueWithoutName2.ToString());
+                TestEnum.ValueWithName1.GetEnumDisplayName().Should().Be(TestEnumNames.Name1);
+                TestEnum.ValueWithName2.GetEnumDisplayName().Should().Be(TestEnumNames.Name2);
+                TestEnum.ValueWithoutName1.GetEnumDisplayName().Should().Be(TestEnum.ValueWithoutName1.ToString());
+                TestEnum.ValueWithoutName2.GetEnumDisplayName().Should().Be(TestEnum.ValueWithoutName2.ToString());
             }
         }
 
@@ -23,16 +24,37 @@ namespace Macabresoft.Core.Tests.Utilities {
         [Category("Unit Tests")]
         public static void GetTypeDisplayName_Should_UseDisplayAttribute() {
             using (new AssertionScope()) {
-                DisplayNameHelper.GetTypeDisplayName(typeof(DisplayNameTestClass)).Should().Be(ClassName);
-                DisplayNameHelper.GetTypeDisplayName(typeof(NonDisplayNameTestClass)).Should().Be(nameof(NonDisplayNameTestClass));
+                typeof(DisplayNameTestClass).GetTypeDisplayName().Should().Be(ClassName);
+                typeof(NonDisplayNameTestClass).GetTypeDisplayName().Should().Be(nameof(NonDisplayNameTestClass));
+            }
+        }
+        
+        [Test]
+        [Category("Unit Tests")]
+        public static void GetPropertyDisplayName_Should_UseDisplayAttribute() {
+            using (new AssertionScope()) {
+                typeof(DisplayNameTestClass).GetPropertyDisplayName(nameof(DisplayNameTestClass.Property1)).Should().Be(Property1Name);
+                typeof(DisplayNameTestClass).GetPropertyDisplayName(nameof(DisplayNameTestClass.Property2)).Should().Be(Property2Name);
+                typeof(NonDisplayNameTestClass).GetPropertyDisplayName(nameof(NonDisplayNameTestClass.Property1)).Should().Be(nameof(NonDisplayNameTestClass.Property1));
+                typeof(NonDisplayNameTestClass).GetPropertyDisplayName(nameof(NonDisplayNameTestClass.Property2)).Should().Be(nameof(NonDisplayNameTestClass.Property2));
             }
         }
 
         [Display(Name = ClassName)]
         private class DisplayNameTestClass {
+            [Display(Name = Property1Name)]
+            public float Property1 { get; }
+            
+            [Display(Name = Property2Name)]
+            public byte Property2 { get; set; }
         }
 
         private class NonDisplayNameTestClass {
+            public object Property1 { get; private set; }
+            
+            public object Property2 {
+                set => Property1 = value;
+            }
         }
     }
 }